+++ VK_Controller/uninstall.sh
@@ -1,1 +1,4 @@
-rm -rf /data/init.log
+#!/system/bin/sh
+rm -rf /cache/init.log
+rm -rf /cache/running.log
+rm -rf ${0%/*}